% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtex.R
\docType{package}
\name{rtex}
\alias{rtex}
\alias{rtex-package}
\title{rtex}
\description{
This package is a tool-box for text mining.
}
\details{
This package helps you see the most frequented words in a word cloud and observe text documents as vectors in a multi-dimensional space.
}
\section{Class TextMiner}{

rtex provides a Reference class named as TextMiner.
This package supports the latest techniques of text mining and sentiment analysis.
It also supports various visualizations for presentation.
}

\author{
Nima Ramezani
}
