% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texer.R
\docType{package}
\name{texer}
\alias{texer}
\alias{texer-package}
\title{texer}
\description{
This package is a tool-box for text mining.
}
\details{
This package helps you see the most frequented words in a word cloud and observe text documents as vectors in a multi-dimensional space.
}
\section{Class TEXT.MINER}{

texer provides a Reference class named as TEXT.MINER.
This package supports the latest techniques of text mining and sentiment analysis.
It also supports various visualizations for presentation.
}

\author{
Nicolas Berta
}
